New Form created.
Added new form for the Orders section, FormOrders_EF. It is related to the model created by following Entity Framework guidelines.

I wanted to point out that the basic form created "FormOrders" had a problem that prevented me to successfully complete CRUD operations related to the Purchase tables. With the newly implemented Entity Framework, and its related form for the Purchase tables "FormOrders_EF", every CRUD operation is successful, so the problem doesn't persist in the current version of the program.