***NEW FEATURES***
Created a new form "FormLogin" with a login page implemented to appear as soon as the program is launched. This new logic is designed for different users with different privileges.

Based on what credentials are received from the login form, a user is allowed or not to be able to see some features of the app more than others. This will depend on what kind of privileges said user possesses; it can be an Administrator, a Data Management Clerk or a simple user that can only View what's inside the database.

If an Administrator is recognized to be connected, it will have the privilege to manage every registered user through a new form I implemented: "FormManageUsers".

Lastly, I developed an encryption procedure for every password inserted. This encryption is a SHA256 type encryption, implemented in a new class called "Utils" in the root folder. Everytime a user logs in, its password will be intercepted and encrypted in SHA256 in order to be confronted with the matching entry in the local database.