*SMALL FIXES*
"FormLogin" has been slightly changed. Buttons "Sign Up" and "Forgot your password?" have been removed. They are not needed for the correct functioning of the application in its current state.

Implemented an event called "tbAny_KeyDown" that mimics the same behavior that "btnLogin_Click" has.
This method lets the user press "Enter" and be able to do everything the click event does. This new event can be performed anywhere, anytime during the user's insertion process in the login form.

This simple Quality of Life improvement has been implemented to facilitate some instinctive action that an average user may perform, like pressing "Enter" key on the keyboard to login instead of hovering on the Login button with the mouse and click it.